@@ -3,14 +3,11 @@ requires = ["setuptools", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [tool.isort]
+profile = "black"
 known_first_party = ["aioredis_cluster"]
 known_third_party = ["aioredis"]
-lines_after_imports = 2
-multi_line_output = 3
-include_trailing_comma = true
-use_parentheses = true
 
 [tool.black]
 line-length = 100
-target-version = ["py36"]
+target-version = ["py38"]
 exclude = '.pyi$'
