@@ -14,8 +14,6 @@ classifiers =
     Development Status :: 5 - Production/Stable
     Programming Language :: Python
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.6
-    Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
@@ -36,9 +34,8 @@ package_dir =
     =src
 install_requires =
     async-timeout
-    dataclasses; python_version < '3.7'
-    hiredis < 3.0.0; platform_python_implementation == "CPython" and python_version < '3.10'
-python_requires = >=3.6.5, <4
+    hiredis < 3.0.0; platform_python_implementation == "CPython"
+python_requires = >=3.8, <4
 zip_safe = False
 include_package_data = True
 
@@ -52,17 +49,17 @@ include =
 devel =
     flake8
     mypy
-    isort>=5.0.0, <6.0.0
+    isort
     mock>=4.0.0
-    black==22.3.0
+    black==22.12.0
     coverage
+    cython
     pytest
     pytest-cov
     pytest-mock
     pytest-asyncio<0.19
     pytest-cov
     pytest-xdist
-    types-dataclasses; python_version < '3.7'
     types-mock
 aioredis =
     aioredis >=1.1.0, <1.4.0; python_version < '3.10'
@@ -70,6 +67,7 @@ aioredis =
 [tool:pytest]
 minversion = 2.9.1
 addopts = --cov-report=term --cov-report=html -v
+asyncio_mode = auto
 
 [coverage:run]
 branch = true
@@ -82,6 +80,7 @@ extend_ignore = W606, E203, E741
 [mypy]
 ignore_missing_imports = True
 check_untyped_defs = True
+no_implicit_optional = False
 
 [mypy-aioredis_cluster._aioredis.*]
 ignore_errors = True
