@@ -4,14 +4,12 @@ from typing import Dict, List, NamedTuple, Optional, Sequence, Set
 
 from aioredis_cluster.abc import AbcPool
 from aioredis_cluster.log import logger
+from aioredis_cluster.pool import POOL_IDLE_CONNECTIOM_TIMEOUT
 from aioredis_cluster.resource_lock import ResourceLock
 from aioredis_cluster.structs import Address
 from aioredis_cluster.typedef import PoolCreator, PoolerBatchCallback
 
-
-__all__ = [
-    "Pooler",
-]
+__all__ = ("Pooler",)
 
 
 @dataclasses.dataclass
@@ -23,10 +21,11 @@ class PoolHolder:
 class PubSubChannel(NamedTuple):
     name: bytes
     is_pattern: bool
+    is_sharded: bool
 
 
 class Pooler:
-    REAP_FREQUENCY = 60.0 * 10
+    REAP_FREQUENCY = POOL_IDLE_CONNECTIOM_TIMEOUT
 
     def __init__(
         self,
@@ -47,7 +46,7 @@ class Pooler:
 
         self._reaper_task = None
         if self._reap_frequency >= 0:
-            loop = asyncio.get_event_loop()
+            loop = asyncio.get_running_loop()
             self._reaper_task = loop.create_task(self._reaper())
 
         self._closed = False
@@ -99,11 +98,18 @@ class Pooler:
     def pools(self) -> List[AbcPool]:
         return [h.pool for h in self._nodes.values()]
 
-    def add_pubsub_channel(self, addr: Address, channel_name: bytes, is_pattern: bool) -> bool:
+    def add_pubsub_channel(
+        self,
+        addr: Address,
+        channel_name: bytes,
+        *,
+        is_pattern: bool = False,
+        is_sharded: bool = False,
+    ) -> bool:
         if addr not in self._pubsub_addrs:
             return False
 
-        channel_obj = PubSubChannel(channel_name, is_pattern)
+        channel_obj = PubSubChannel(channel_name, is_pattern, is_sharded)
         if channel_obj in self._pubsub_channels:
             return False
 
@@ -111,8 +117,14 @@ class Pooler:
         self._pubsub_addrs[addr].add(channel_obj)
         return True
 
-    def remove_pubsub_channel(self, channel_name: bytes, is_pattern: bool) -> bool:
-        channel_obj = PubSubChannel(channel_name, is_pattern)
+    def remove_pubsub_channel(
+        self,
+        channel_name: bytes,
+        *,
+        is_pattern: bool = False,
+        is_sharded: bool = False,
+    ) -> bool:
+        channel_obj = PubSubChannel(channel_name, is_pattern, is_sharded)
         addr = self._pubsub_channels.pop(channel_obj, None)
         if addr:
             if addr in self._pubsub_addrs:
@@ -120,8 +132,14 @@ class Pooler:
             return True
         return False
 
-    def get_pubsub_addr(self, channel_name: bytes, is_pattern: bool) -> Optional[Address]:
-        channel_obj = PubSubChannel(channel_name, is_pattern)
+    def get_pubsub_addr(
+        self,
+        channel_name: bytes,
+        *,
+        is_pattern: bool = False,
+        is_sharded: bool = False,
+    ) -> Optional[Address]:
+        channel_obj = PubSubChannel(channel_name, is_pattern, is_sharded)
         return self._pubsub_channels.get(channel_obj)
 
     async def close(self) -> None:
