@@ -1,11 +1,7 @@
 from typing import Awaitable, Callable
 
 from aioredis_cluster.aioredis.util import _NOTSET, wait_convert, wait_ok
-from aioredis_cluster.util import (
-    parse_cluster_nodes,
-    parse_cluster_slaves,
-    parse_info,
-)
+from aioredis_cluster.util import parse_cluster_nodes, parse_cluster_slaves, parse_info
 
 
 class ClusterCommandsMixin:
