@@ -9,7 +9,7 @@ from aioredis_cluster.aioredis.util import _NOTSET
 
 from .cluster import ClusterCommandsMixin
 from .custom import StreamCustomCommandsMixin
-
+from .sharded_pubsub import ShardedPubSubCommandsMixin
 
 __all__ = (
     "Redis",
@@ -35,7 +35,9 @@ def blocked_for_cluster(method: Callable[..., _T]) -> Callable[..., _T]:
     return wrapper
 
 
-class RedisCluster(ClusterCommandsMixin, StreamCustomCommandsMixin, Redis):
+class RedisCluster(
+    ClusterCommandsMixin, StreamCustomCommandsMixin, ShardedPubSubCommandsMixin, Redis
+):
     # proxy methods to connection (cluster instance)
     async def all_masters(self) -> List[Redis]:
         self._only_for_cluster()
